requirements.txt
torch
torchvision
torchaudio
matplotlib
numpy
pillow
scikit-learn
tqdm
flask 
opencv-python